+++ doc/CHANGELOG.md
@@ -2,11 +2,14 @@
 
 
 ## [2024 mmm dd] v0.4.2
+### Documentation
+- [ ] add a contributing guide
 ### Features
 - [x] make minimum distance to boundary for interior points of a `MeshCell` adjustable with `set_interior_point_tolerance()` method
 - [ ] make computation of interior gradients optional
+- [x] add methods to `MeshCell` to get unit tangent and unit normal vectors, and the derivative norm
 ### Maintenance
-- [ ] vectorize interior value computation
+- [x] vectorize interior value computation
 ### Tests
 - [x] add ghost cell to mesh cell testing library
 - [x] add test for interior points